requirements.txt
torch
transformers
pymupdf
python-docx
nltk
langdetect
fastapi
uvicorn
requests
PyPDF2
aiosqlite
huggingface_hub
python-multipart
tenacity
apscheduler